add day3

+++ src/day1.rs
@@ -2,10 +2,10 @@ use std::collections::HashSet;
 
 use itertools::Itertools;
 
-pub fn day1_part1(input: String) -> u32 {
+pub fn day1_part1(input: String) -> u64 {
     let values = input
         .split_whitespace()
-        .map(|v| v.parse::<u32>().unwrap())
+        .map(|v| v.parse::<u64>().unwrap())
         .collect::<HashSet<_>>();
 
     let complements = values
@@ -17,10 +17,10 @@ pub fn day1_part1(input: String) -> u32 {
     m * (2020 - m)
 }
 
-pub fn day1_part2(input: String) -> u32 {
+pub fn day1_part2(input: String) -> u64 {
     input
         .split_whitespace()
-        .map(|v| v.parse::<u32>().unwrap())
+        .map(|v| v.parse::<u64>().unwrap())
         .combinations(3)
         .find(|vals| vals[0] + vals[1] + vals[2] == 2020)
         .map(|vals| vals[0] * vals[1] * vals[2])
